+++ access_manager.go
@@ -1,7 +1,10 @@
 // Package restrict provides an authorization library, with a hybrid of RBAC and ABAC models.
 package restrict
 
-import "fmt"
+import (
+	"fmt"
+	"github.com/el-mike/restrict/internal/utils"
+)
 
 // AccessManager - an entity responsible for checking the authorization. It uses underlying
 // PolicyProvider to test an AccessRequest against currently used PolicyDefinition.
@@ -31,30 +34,31 @@ func (am *AccessManager) Authorize(request *AccessRequest) error {
 		return newRequestMalformedError(request, fmt.Errorf("Missing roles or resourceName"))
 	}
 
-	accessErrors := PermissionErrors{}
+	allPermissionErrors := PermissionErrors{}
 
 	for _, roleName := range roles {
-		permissionDeniedError, err := am.authorize(request, roleName, resourceName, []string{})
+		permissionErrors, err := am.authorize(request, roleName, resourceName, []string{})
 
+		// If error is not authorization-specific, we return immediately.
 		if err != nil {
 			return err
 		}
 
 		// If AccessRequest is satisfied by a Role, we return immediately.
-		if permissionDeniedError == nil {
+		if permissionErrors == nil {
 			return nil
 		}
 
 		// Otherwise, we save it to PermissionErrors, so we can return it to the caller
 		// if no Role satisfies the AccessRequest.
-		accessErrors = append(accessErrors, permissionDeniedError)
+		allPermissionErrors = append(allPermissionErrors, permissionErrors...)
 	}
 
-	return newAccessDeniedError(request, accessErrors)
+	return newAccessDeniedError(request, allPermissionErrors)
 }
 
 // authorize - helper function for decoupling role and resource names retrieval from recursive search.
-func (am *AccessManager) authorize(request *AccessRequest, roleName, resourceName string, checkedRoles []string) (*PermissionError, error) {
+func (am *AccessManager) authorize(request *AccessRequest, roleName, resourceName string, checkedRoles []string) (PermissionErrors, error) {
 	role, err := am.policyManager.GetRole(roleName)
 	if err != nil {
 		return nil, err
@@ -69,17 +73,22 @@ func (am *AccessManager) authorize(request *AccessRequest, roleName, resourceNam
 	}
 
 	parents := role.Parents
+	allPermissionErrors := PermissionErrors{}
 
 	for _, action := range request.Actions {
 		if action == "" {
-			return nil, newRequestMalformedError(request, fmt.Errorf("Action cannot be empty"))
+			return nil, newRequestMalformedError(request, fmt.Errorf("action cannot be empty"))
 		}
 
-		validationError := am.validateAction(grants, action, roleName, request)
+		permissionErrors, err := am.validateAction(grants, action, roleName, request)
+		// If non-policy related error happened, we return it directly.
+		if err != nil {
+			return nil, err
+		}
 
 		// If access is not granted for given action on current Role, check if
 		// any parent Role can satisfy the request.
-		if validationError != nil && len(parents) > 0 {
+		if permissionErrors != nil && len(parents) > 0 {
 			checkedRoles = append(checkedRoles, roleName)
 
 			for _, parent := range parents {
@@ -92,101 +101,119 @@ func (am *AccessManager) authorize(request *AccessRequest, roleName, resourceNam
 
 				// If parent has already been checked, we want to return an error - otherwise
 				// this function will fall into infinite loop.
-				if am.isRoleChecked(parent, checkedRoles) {
+				if utils.StringSliceContains(checkedRoles, parent) {
 					return nil, newRoleInheritanceCycleError(checkedRoles)
 				}
 
-				permissionDeniedError, err := am.authorize(parentRequest, parent, resourceName, checkedRoles)
-
-				// If the returned error is not an access error (meaning something not related
-				// to Policy validation happened), we want to return the actual error to the caller.
+				parentPermissionErrors, err := am.authorize(parentRequest, parent, resourceName, checkedRoles)
 				if err != nil {
 					return nil, err
 				}
 
 				// If .authorize call with parent Role has returned nil,
 				// that means the request is satisfied.
-				// Otherwise, returned error should not override the original one.
-				if permissionDeniedError == nil {
-					validationError = nil
+				// Otherwise, returned errors should not override the original ones.
+				if parentPermissionErrors == nil {
+					permissionErrors = nil
 				}
 			}
 		}
 
 		// If request has not been granted, abort the loop and return an error,
 		// skipping rest of the Actions in the request.
-		if validationError != nil {
-			if permissionDeniedError, ok := validationError.(*PermissionError); ok {
-				return permissionDeniedError, nil
-			}
+		if permissionErrors != nil {
+			allPermissionErrors = append(allPermissionErrors, permissionErrors...)
 
-			// If the error is not PermissionError, error was not specific to Policy validation.
-			return nil, validationError
+			// If CompleteValidation is false, we want to return early.
+			// Otherwise, loop will continue for the rest of the actions.
+			if !request.CompleteValidation {
+				return allPermissionErrors, nil
+			}
 		}
+	}
+
+	if len(allPermissionErrors) > 0 {
+		return allPermissionErrors, nil
 	}
 
 	return nil, nil
 }
 
 // validateAction - checks whether a Permission is granted for Action.
-func (am *AccessManager) validateAction(permissions []*Permission, action, roleName string, request *AccessRequest) error {
-	var conditionsCheckError error
+func (am *AccessManager) validateAction(permissions []*Permission, action, roleName string, request *AccessRequest) (PermissionErrors, error) {
+	permissionErrors := PermissionErrors{}
 
 	for _, permission := range permissions {
 		if permission.Action == action {
+			// If a Permission with given Action is found, and has no Conditions, access should be granted.
 			if len(permission.Conditions) == 0 || request.SkipConditions {
-				return nil
+				return nil, nil
 			}
 
-			conditionsCheckError = am.checkConditions(permission, request)
-
-			// If error is nil, Conditions have been satisfied. Otherwise, we don't want to
-			// break the loop, as other Permissions can have the same action.
-			if conditionsCheckError == nil {
-				return nil
+			conditionErrors, err := am.checkConditions(permission, request)
+			// If non-policy related error happened, we return it directly.
+			if err != nil {
+				return nil, err
 			}
+
+			// If error is nil, Conditions have been satisfied.
+			if conditionErrors == nil {
+				return nil, nil
+			}
+
+			// Otherwise, we add new PermissionError to result slice.
+			permissionError := newPermissionError(action, roleName, request.Resource.GetResourceName(), conditionErrors)
+			permissionErrors = append(permissionErrors, permissionError)
 		}
 	}
 
-	// If condition error is not related to Policy validation, return it directly.
-	// Otherwise, we wrap it in a PermissionError.
-	if conditionsCheckError != nil && !am.isAccessError(conditionsCheckError) {
-		return conditionsCheckError
+	// If there are no permissionErrors at this point, this means there was no Permission
+	// for given Action (even one with failing Conditions).
+	// In such case, a PermissionError with no ConditionErrors is added to reflect that.
+	if len(permissionErrors) == 0 {
+		permissionError := newPermissionError(action, roleName, request.Resource.GetResourceName(), nil)
+		permissionErrors = append(permissionErrors, permissionError)
 	}
 
-	return newPermissionError(action, roleName, request.Resource.GetResourceName(), conditionsCheckError)
+	return permissionErrors, nil
 }
 
 // checkConditions - returns nil if all conditions specified for given actions
 // are satisfied, error otherwise.
-func (am *AccessManager) checkConditions(permission *Permission, request *AccessRequest) error {
+func (am *AccessManager) checkConditions(permission *Permission, request *AccessRequest) (ConditionErrors, error) {
 	if permission.Conditions == nil {
-		return nil
+		return nil, nil
 	}
+
+	conditionErrors := ConditionErrors{}
 
 	for _, condition := range permission.Conditions {
 		if err := condition.Check(request); err != nil {
-			return err
+			// If error returned is ConditionNotSatisfiedError, we add it to the result slice.
+			// Otherwise, we want to abort immediately and return it directly.
+			if conditionError, ok := err.(*ConditionNotSatisfiedError); ok {
+				conditionErrors = append(conditionErrors, conditionError)
+
+				// If CompleteValidation is not enabled, we return first encountered error.
+				if !request.CompleteValidation {
+					break
+				}
+			} else {
+				return nil, err
+			}
 		}
 	}
 
-	return nil
-}
-
-// isRoleChecked - returns true if role is in parents slice, false otherwise.
-func (am *AccessManager) isRoleChecked(role string, parents []string) bool {
-	for _, parent := range parents {
-		if role == parent {
-			return true
-		}
+	if len(conditionErrors) > 0 {
+		return conditionErrors, nil
 	}
 
-	return false
+	return nil, nil
 }
 
-// isAccessError - returns true if error is caused by reasons related to Policy validation
+// IsAccessError - returns true if error is caused by reasons related to Policy validation
 // (Conditions checks, missing Permissions etc.), false otherwise (malformed Policy, reflection errors, etc.)
-func (am *AccessManager) isAccessError(err error) bool {
+func (am *AccessManager) IsAccessError(err error) bool {
 	switch err.(type) {
 	case *ConditionNotSatisfiedError,
 		*PermissionError,